Updated circleci config.yaml. Retrieve version from artifact.
CircleCI may fill the CIRCLE_TAG only if initially tags are pushed instead of tagless-commits. This is too easy to forget to do, therefore retrieve the version from the artifact and stick with the filter when to run the release job.

See https://support.circleci.com/hc/en-us/articles/360020342494-The-built-in-environment-variable-CIRCLE-TAG-is-not-present